+++ examples/java/com/ibm/watson/developer_cloud/document_conversion/v1/DocumentConversionExample.java
@@ -1,15 +1,14 @@
 /**
  * Copyright 2015 IBM Corp. All Rights Reserved.
  *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
  */
 package com.ibm.watson.developer_cloud.document_conversion.v1;
 
@@ -18,7 +17,7 @@ import java.io.File;
 import com.ibm.watson.developer_cloud.document_conversion.v1.model.Answers;
 
 /**
- * Example class that shows the various usage scenario's of the Document Conversion service
+ * Example class that shows the various usage scenario's of the Document Conversion service.
  */
 public class DocumentConversionExample {
 
@@ -27,8 +26,8 @@ public class DocumentConversionExample {
     service.setUsernameAndPassword("<username>", "<password>");
 
     File html = new File("src/test/resources/document_conversion/html-with-extra-content-input.htm");
-    File pdf =  new File("src/test/resources/document_conversion/pdf-with-sections-input.pdf");
-    File doc =  new File("src/test/resources/document_conversion/word-document-heading-input.doc");
+    File pdf = new File("src/test/resources/document_conversion/pdf-with-sections-input.pdf");
+    File doc = new File("src/test/resources/document_conversion/word-document-heading-input.doc");
 
     System.out.println("Convert html document to Answers");
     Answers htmlToAnswers = service.convertDocumentToAnswer(html).execute();