+++ src/main/java/com/ibm/watson/developer_cloud/alchemy/v1/util/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Utility classes to create API requests and transform JSON into POJO classes among other things
- */
-package com.ibm.watson.developer_cloud.alchemy.v1.util;
-