+++ src/main/java/com/ibm/watson/developer_cloud/personality_insights/v2/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Personality Insights Service v1
- */
-package com.ibm.watson.developer_cloud.personality_insights.v2;
-