+++ src/main/java/com/ibm/watson/developer_cloud/retrieve_and_rank/v1/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Retrieve and Rank Service v1
- */
-package com.ibm.watson.developer_cloud.retrieve_and_rank.v1;
-