+++ src/main/java/com/ibm/watson/developer_cloud/service/package-info.java
@@ -1,6 +1,0 @@
-/**
- * This package contains Watson and Alchemy generic service implementation.<br>
- * Services will extend this classes with concrete implementations
- */
-package com.ibm.watson.developer_cloud.service;
-