+++ src/main/java/com/ibm/watson/developer_cloud/speech_to_text/v1/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Speech to Text Service v1
- */
-package com.ibm.watson.developer_cloud.speech_to_text.v1;
-