+++ src/main/java/com/ibm/watson/developer_cloud/tradeoff_analytics/v1/model/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Classes that map the HTTP service requests and responses into POJOs
- */
-package com.ibm.watson.developer_cloud.tradeoff_analytics.v1.model;
-