+++ src/main/java/com/ibm/watson/developer_cloud/util/package-info.java
@@ -1,5 +1,0 @@
-/**
- * This package contains utility classes for working with HTTP requests and responses.
- */
-package com.ibm.watson.developer_cloud.util;
-